requirements.txt
openai == 1.7.1
python-dotenv == 1.0.0
langchain-community == 0.0.11
langchain == 0.1.0
streamlit
pymongo
langchain-openai
ctransformers
transformers
torch
pypdf
sentence-transformers
gpt4all
faiss-cpu
lxml
requests
bs4
tiktoken
langchain-mongodb
requests_html
